# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "春季nodejs_finaly_API文件",
    "description": "示範撰寫API文件",
    "version": "1.0.0"
  },
  "host": "localhost:3005",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
    "/posts": {
      "get": {
        "tags": [
          "取得"
        ],
        "description": "取得所有貼文",
        "parameters": [
          {
            "name": "timeSort",
            "in": "query",
            "type": "string"
          },
          {
            "name": "q",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "所有貼文資訊",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 200
                },
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "string",
                        "example": "文章id"
                      },
                      "user": {
                        "type": "object",
                        "properties": {
                          "name": {
                            "type": "string",
                            "example": "作者名稱"
                          },
                          "photo": {
                            "type": "string",
                            "example": "作者頭像"
                          }
                        }
                      },
                      "image": {
                        "type": "string",
                        "example": "文章圖片"
                      },
                      "content": {
                        "type": "string",
                        "example": "文章內容"
                      },
                      "likes": {
                        "type": "string",
                        "example": "讚數"
                      },
                      "type": {
                        "type": "string",
                        "example": "文章類型"
                      },
                      "tags": {
                        "type": "array",
                        "example": [
                          "標籤1",
                          "標籤2"
                        ],
                        "items": {
                          "type": "string"
                        }
                      },
                      "createdAt": {
                        "type": "string",
                        "example": "發文時間"
                      },
                      "comments": {
                        "type": "array",
                        "example": [
                          "留言1",
                          "留言2"
                        ],
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "刪除"
        ],
        "description": "刪除個人所有貼文",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "使用者認證",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "資料格式",
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "type": "string",
                  "example": "使用者密碼"
                }
              },
              "required": [
                "password"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "刪除資訊",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 200
                },
                "message": {
                  "type": "string",
                  "example": "全部清空"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/post/user/{userId}": {
      "get": {
        "tags": [
          "取得"
        ],
        "description": "取得個人所有貼文",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "使用者認證",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "個人所有貼文資訊",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 200
                },
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "string",
                        "example": "文章id"
                      },
                      "user": {
                        "type": "string",
                        "example": "使用者id"
                      },
                      "image": {
                        "type": "string",
                        "example": "文章圖片"
                      },
                      "content": {
                        "type": "string",
                        "example": "文章內容"
                      },
                      "likes": {
                        "type": "string",
                        "example": "讚數"
                      },
                      "type": {
                        "type": "string",
                        "example": "文章類型"
                      },
                      "tags": {
                        "type": "array",
                        "example": [
                          "標籤1",
                          "標籤2"
                        ],
                        "items": {
                          "type": "string"
                        }
                      },
                      "createdAt": {
                        "type": "string",
                        "example": "發文時間"
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/posts/{postId}": {
      "get": {
        "tags": [],
        "description": "查看單一貼文",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "單一貼文資訊",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 200
                },
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "string",
                        "example": "文章id"
                      },
                      "user": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string",
                            "example": "發文者id"
                          },
                          "name": {
                            "type": "string",
                            "example": "發文者名稱"
                          },
                          "photo": {
                            "type": "string",
                            "example": "發文者頭像"
                          }
                        }
                      },
                      "image": {
                        "type": "string",
                        "example": "文章圖片"
                      },
                      "content": {
                        "type": "string",
                        "example": "文章內容"
                      },
                      "likes": {
                        "type": "string",
                        "example": "讚數"
                      },
                      "type": {
                        "type": "string",
                        "example": "文章類型"
                      },
                      "tags": {
                        "type": "array",
                        "example": [
                          "標籤1",
                          "標籤2"
                        ],
                        "items": {
                          "type": "string"
                        }
                      },
                      "createdAt": {
                        "type": "string",
                        "example": "發文時間"
                      },
                      "comments": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "_id": {
                              "type": "string",
                              "example": "留言id"
                            },
                            "comment": {
                              "type": "string",
                              "example": "留言內容"
                            },
                            "user": {
                              "type": "object",
                              "properties": {
                                "name": {
                                  "type": "string",
                                  "example": "留言者名稱"
                                }
                              }
                            },
                            "post": {
                              "type": "string",
                              "example": "留言貼文id"
                            },
                            "createdAt": {
                              "type": "string",
                              "example": "留言時間"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/post": {
      "post": {
        "tags": [
          "新增"
        ],
        "description": "使用者發文",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "資料格式",
            "schema": {
              "type": "object",
              "properties": {
                "user": {
                  "type": "string",
                  "example": "使用者id"
                },
                "content": {
                  "type": "string",
                  "example": "文章內容"
                },
                "type": {
                  "type": "string",
                  "example": "文章類型"
                },
                "image": {
                  "type": "string",
                  "example": "文章圖片"
                },
                "tags": {
                  "type": "array",
                  "example": [
                    "標籤1",
                    "標籤2"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              },
              "required": [
                "user",
                "content",
                "type"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "發文資訊",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 200
                },
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "user": {
                        "type": "string",
                        "example": "使用者id"
                      },
                      "content": {
                        "type": "string",
                        "example": "文章內容"
                      },
                      "type": {
                        "type": "string",
                        "example": "文章類型"
                      },
                      "image": {
                        "type": "string",
                        "example": "文章圖片"
                      },
                      "tags": {
                        "type": "array",
                        "example": [
                          "標籤1",
                          "標籤2"
                        ],
                        "items": {
                          "type": "string"
                        }
                      },
                      "likes": {
                        "type": "number",
                        "example": 0
                      },
                      "_id": {
                        "type": "string",
                        "example": "文章id"
                      },
                      "createdAt": {
                        "type": "string",
                        "example": "發文時間"
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/posts/{postId}/comment": {
      "post": {
        "tags": [
          "新增"
        ],
        "description": "新增貼文留言",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "使用者認證",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "comment": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "新增留言資訊",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 200
                },
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "comment": {
                        "type": "string",
                        "example": "留言內容"
                      },
                      "user": {
                        "type": "string",
                        "example": "留言者id"
                      },
                      "post": {
                        "type": "string",
                        "example": "留言貼文id"
                      },
                      "_id": {
                        "type": "string",
                        "example": "留言id"
                      },
                      "createdAt": {
                        "type": "string",
                        "example": "留言時間"
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/posts/{postId}/like": {
      "post": {
        "tags": [
          "新增"
        ],
        "description": "按讚貼文",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "使用者認證",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "按讚",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 200
                },
                "message": {
                  "type": "string",
                  "example": "按讚成功"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/posts/{postId}/unlike": {
      "delete": {
        "tags": [
          "刪除"
        ],
        "description": "取消按讚貼文",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "使用者認證",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "取消按讚",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 200
                },
                "message": {
                  "type": "string",
                  "example": "取消按讚成功"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/post/{postId}": {
      "patch": {
        "tags": [
          "修改"
        ],
        "description": "修改個人貼文",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "使用者認證",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "資料格式",
            "schema": {
              "type": "object",
              "properties": {
                "content": {
                  "type": "string",
                  "example": "文章內容"
                },
                "type": {
                  "type": "string",
                  "example": "文章類型"
                },
                "image": {
                  "type": "string",
                  "example": "文章圖片"
                },
                "tags": {
                  "type": "array",
                  "example": [
                    "標籤1",
                    "標籤2"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              },
              "required": [
                "content",
                "type"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "更新資訊",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 200
                },
                "message": {
                  "type": "string",
                  "example": "更新成功"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "刪除"
        ],
        "description": "刪除個人貼文",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "使用者認證",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "刪除資訊",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 200
                },
                "message": {
                  "type": "string",
                  "example": "刪除成功"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  }
}